+++ src/terminus.vala
@@ -14,12 +14,13 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>. */
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
 
 using Gtk;
 using Gee;
 
-//project version = 1.5.0
+//project version = 1.7.1
 
 namespace Terminus {
 	TerminusRoot     main_root;
@@ -206,7 +207,7 @@ namespace Terminus {
 		private Terminus.Window ? guake_window;
 		private bool ready;
 		private int extcall;
-		private bool guake_has_focus;
+		private string ? guake_title;
 
 		private bool tmp_launch_terminal;
 		private bool tmp_launch_guake;
@@ -221,7 +222,7 @@ namespace Terminus {
 			main_root            = this;
 			this.guake_terminal  = null;
 			this.guake_window    = null;
-			this.guake_has_focus = false;
+			this.guake_title     = null;
 
 			bool binded_key = Terminus.bindkey.set_bindkey(Terminus.keybind_settings.get_string("guake-mode"));
 
@@ -229,6 +230,7 @@ namespace Terminus {
 
 			this.launch_guake = parameters.launch_guake;
 			this.check_guake  = parameters.check_guake;
+			this.guake_title  = parameters.UUID;
 
 			this.tmp_launch_terminal = true;
 			this.tmp_launch_guake    = false;
@@ -278,6 +280,10 @@ namespace Terminus {
 					    show_hide_global(this.extcall);
 					}
 				}, () => {
+					// if we are checking to launch guake, and there is already a process, return an error
+					if (this.check_guake) {
+						Posix.exit(1);
+					}
 					// if there is another Terminus process, ask it to open a new window and exit
 					RemoteControlInterface server = Bus.get_proxy_sync(BusType.SESSION, "com.rastersoft.terminus", "/com/rastersoft/terminus");
 					if (this.tmp_launch_terminal) {
@@ -336,16 +342,6 @@ namespace Terminus {
 			Terminus.bindkey.show_guake.connect(this.show_hide);
 		}
 
-		public bool focus_in(Gdk.EventFocus event) {
-			this.guake_has_focus = true;
-			return false;
-		}
-
-		public bool focus_out(Gdk.EventFocus event) {
-			this.guake_has_focus = false;
-			return false;
-		}
-
 		public int create_window(bool guake_mode) {
 			Terminus.Window window;
 
@@ -353,11 +349,9 @@ namespace Terminus {
 				if (this.guake_terminal == null) {
 					this.guake_terminal = new Terminus.Base();
 				}
-				window            = new Terminus.Window(true, tid_counter, this.guake_terminal);
+				window            = new Terminus.Window(true, tid_counter, this.guake_terminal, this.guake_title);
 				this.guake_window = window;
 				Terminus.bindkey.show_guake.connect(this.show_hide);
-				this.guake_window.focus_in_event.connect(this.focus_in);
-				this.guake_window.focus_out_event.connect(this.focus_out);
 			} else {
 				window = new Terminus.Window(false, tid_counter);
 			}
@@ -421,12 +415,7 @@ namespace Terminus {
 
 			// mode 2
 			if (this.guake_window.visible) {
-				if ((check_wayland() != 0) && (!this.guake_has_focus)) {
-					this.guake_window.hide();
-					this.guake_window.show();
-				} else {
-					this.guake_window.hide();
-				}
+				this.guake_window.hide();
 			} else {
 				this.guake_window.present();
 			}